Add GodotNode to prevent orphaned nodes
Also update namespaces
