+++ Binding/ControlBinders/OptionButtonControlBinder.cs
@@ -1,10 +1,10 @@
-using ControlBinding.Collections;
-using ControlBinding.EventArgs;
+using Godot.Community.ControlBinding.Collections;
+using Godot.Community.ControlBinding.EventArgs;
 using Godot;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace ControlBinding.ControlBinders;
+namespace Godot.Community.ControlBinding.ControlBinders;
 public partial class OptionButtonControlBinder : ControlBinderBase
 {
     private readonly List<string> _allowedTwoBindingProperties = new List<string>(){
