+++ Binding/ControlBinders/RangeControlBinder.cs
@@ -1,10 +1,10 @@
-using ControlBinding.EventArgs;
+using Godot.Community.ControlBinding.EventArgs;
 using Godot;
 using System;
 using System.Collections.Generic;
 using Range = Godot.Range;
 
-namespace ControlBinding.ControlBinders;
+namespace Godot.Community.ControlBinding.ControlBinders;
 
 public partial class RangeControlBinder : ControlBinderBase
 {
