+++ Binding/EventHandlers/PropertyChangedEventHandler.cs
@@ -1,0 +1,9 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace Godot.Community.ControlBinding
+{
+    public delegate void PropertyChangedEventHandler(GodotObject owner, string propertyName);
+}
