+++ Binding/Formatters/SceneFormatter.cs
@@ -1,8 +1,8 @@
 using System;
-using ControlBinding.Utilities;
+using Godot.Community.ControlBinding.Utilities;
 using Godot;
 
-namespace ControlBinding.Formatters
+namespace Godot.Community.ControlBinding.Formatters
 {
     public class SceneFormatter : ISceneFormatter
     {
