+++ Binding/Interfaces/IControlBinder.cs
@@ -1,6 +1,6 @@
-using ControlBinding.EventArgs;
+using Godot.Community.ControlBinding.EventArgs;
 
-namespace ControlBinding.ControlBinders;
+namespace Godot.Community.ControlBinding.ControlBinders;
 public interface IControlBinder
 {
     bool CanBindFor(System.Object control);
