+++ Binding/Interfaces/IObservableNode.cs
@@ -1,0 +1,12 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace Godot.Community.ControlBinding.Interfaces
+{
+    public interface IObservableNode
+    {
+        void SetViewModelData(object viewModelData);
+    }
+}
