+++ Binding/Interfaces/IObservableObject.cs
@@ -1,7 +1,10 @@
-namespace ControlBinding.Interfaces;
+using System;
+using Godot.Community.ControlBinding;
+
+namespace Godot.Community.ControlBinding.Interfaces;
 
 internal interface IObservableObject
 {
     void OnPropertyChanged(string name);
-    void SetViewModelData(object viewModelData);
+    event PropertyChangedEventHandler PropertyChanged; 
 }
